Add devtools panel
