+++ packages/query-devtools/src/contexts/QueryDevtoolsContext.ts
@@ -1,5 +1,4 @@
 import { createContext, useContext } from 'solid-js'
-import type { Accessor } from 'solid-js'
 import type { Query, QueryClient, onlineManager } from '@tanstack/query-core'
 
 type XPosition = 'left' | 'right'
@@ -29,6 +28,7 @@ export interface QueryDevtoolsProps {
   initialIsOpen?: boolean
   errorTypes?: Array<DevtoolsErrorType>
   shadowDOMTarget?: ShadowRoot
+  onClose?: () => unknown
 }
 
 export const QueryDevtoolsContext = createContext<QueryDevtoolsProps>({
@@ -41,12 +41,4 @@ export const QueryDevtoolsContext = createContext<QueryDevtoolsProps>({
 
 export function useQueryDevtoolsContext() {
   return useContext(QueryDevtoolsContext)
-}
-
-export const ThemeContext = createContext<Accessor<'light' | 'dark'>>(
-  () => 'dark' as const,
-)
-
-export function useTheme() {
-  return useContext(ThemeContext)
 }
