+++ packages/react-query-devtools/src/production.ts
@@ -3,3 +3,4 @@
 import * as devtools from './devtools'
 
 export const ReactQueryDevtools = devtools.ReactQueryDevtools
+export const ReactQueryDevtoolsPanel = devtools.ReactQueryDevtoolsPanel
